STYLE: 엔티티 타입 변경

+++ backend/src/main/java/com/showing/backend/db/entity/Tag.java
@@ -24,5 +24,5 @@ public class Tag extends BaseTimeEntity {
     private String tagName;
 
     @Column
-    private Double weight;
+    private Float weight;
 }
